Pause Screen
Created a basic pause screen with the options to change volume level, resume game and exit to main menu. Pause can be accessed with the ESC key and the ESC key can also be used to exit the screen/resume the game. Currently it is in the settings screen but will later be transferred to the game scene instead.